Streamlit App/requirements.txt
torch
torchvision
numpy
pillow
streamlit
opencv-python
matplotlib
